[Sher] optimize code
