+++ src/services/publicdata/chainInfo.ts
@@ -67,11 +67,11 @@ export async function getBlockInfo() {
 }
 
 export async function getTxInfo(tx_id: any) {
-    let baseURL = sidecarURLKrypton;
+    let baseURL = sidecarURLXenon;
     switch (getNetworkFromStorage()) {
-        case "krypton": baseURL = sidecarURLKrypton; break;
-        case "xenon": baseURL = sidecarURLXenon; break;
-        case "mainnet": break; //TODO
+        case "Krypton": baseURL = sidecarURLKrypton; break;
+        case "Xenon": baseURL = sidecarURLXenon; break;
+        case "Mainnet": break; //TODO
         default: break;
     }
     return request(`${baseURL}/v1/tx/${tx_id}`, {
